backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic-settings==2.2.1
boto3==1.34.120
celery==5.4.0
redis==5.0.4
pymupdf==1.24.3
python-multipart==0.0.9
qdrant-client>=1.8.0
sentence-transformers==2.7.0
torch==2.3.0
google-generativeai==0.5.1
